third_party/WebKit/LayoutTests/platform/win/fast/forms/select-popup/popup-menu-position-expected.txt


PASS popupWindowRect.x is 0
PASS popupWindowRect.y is menuHeight
PASS popupWindowRect.x is screenWidth - popupWindowRect.width
PASS popupWindowRect.y is screenHeight - popupWindowRect.height - menuHeight
PASS popupWindowRect.x is screenWidth - 100
PASS popupWindowRect.y is menuHeight
PASS popupWindowRect.x is 0
PASS popupWindowRect.y is menuHeight
PASS popupWindowRect.x is screenWidth - 250 - 10
PASS popupWindowRect.y is menuHeight
PASS popupWindowRect.x is 0
PASS popupWindowRect.y is screenHeight - 100 - popupWindowRect.height
PASS popupWindowRect.x is 100 - menuWidth * 0.1
PASS popupWindowRect.y is 100 + menuHeight * 1.1
PASS successfullyParsed is true

TEST COMPLETE
